LPD-28349 Move createSitePage function to portal.ts file
