+++ modules/test/playwright/tests/analytics-settings-web/utils/analyticsSettings.ts
@@ -19,37 +19,6 @@ export async function connectToAnalyticsCloud(page: Page) {
 	await page.getByRole('button', {name: 'Connect'}).click();
 }
 
-export const createSitePage = async function ({
-	apiHelpers,
-	pageTitle,
-	siteName = 'Guest',
-}: {
-	apiHelpers: ApiHelpers;
-	pageTitle: string;
-	siteName?: string;
-}) {
-	const company =
-		await apiHelpers.jsonWebServicesCompany.getCompanyByWebId(
-			'liferay.com'
-		);
-
-	const group = await apiHelpers.jsonWebServicesGroup.getGroupByKey(
-		company.companyId,
-		siteName
-	);
-
-	return await apiHelpers.headlessDelivery.createSitePage({
-		pageDefinition: getPageDefinition([
-			getFragmentDefinition({
-				id: getRandomString(),
-				key: 'BASIC_COMPONENT-heading',
-			}),
-		]),
-		siteId: group.groupId,
-		title: pageTitle,
-	});
-};
-
 export async function disconnectFromAnalyticsCloud(page: Page) {
 	const disconnectButton = page.getByRole('button', {name: 'Disconnect'});
