+++ modules/test/playwright/tests/osb-faro-web/page.spec.ts
@@ -13,8 +13,6 @@ import {loginTest} from '../../fixtures/loginTest';
 import {liferayConfig} from '../../liferay.config';
 import getRandomString from '../../utils/getRandomString';
 import {
-	createSitePage,
-	navigateToSitePage,
 	syncAnalyticsCloud,
 } from '../analytics-settings-web/utils/analyticsSettings';
 import {createChannel, switchChannel} from './utils/channel';
@@ -26,6 +24,7 @@ import {
 } from './utils/navigation';
 import {changeTimeFilter} from './utils/time-filter';
 import {expectNotToBeVisible, expectToBeVisible} from './utils/utils';
+import {createSitePage, navigateToSitePage} from './utils/portal';
 
 export const test = mergeTests(
 	apiHelpersTest,
